Cleanup

+++ src/IceRpc/Slice/SliceDecoder.cs
@@ -986,7 +986,7 @@ public ref partial struct SliceDecoder
                         serverAddress = new ServerAddress(
                             Protocol.Ice,
                             host: "opaque", // not a real host obviously
-                            port: (ushort)Protocol.Ice.DefaultPort,
+                            port: Protocol.Ice.DefaultPort,
                             TransportNames.Opaque,
                             builder.ToImmutable());
                         break;
